requirements.txt
splat64[mips]

